Release version v0.8.299
Metacello new
    baseline: 'GToolkit';
    repository: 'github://feenkcom/gtoolkit:v0.8.299/src';
    load
All commits (inlcuding upstream repositories) since last build:
https://github.com/feenkcom/Bloc/commit/6eaaed by Alistair Grant
Avoid `CollectionIsEmpty: a ProcessList() is empty` exceptions

The exception is triggered in BlHostPulseLoop>>wait when it checks the
priority of the nextReadyProcess. The problem is that nextReadyProcess
checks that the list is not empty and then tries to retrieve the entry,
however the process can be interrupted by a higher priority process
(which also typically moves the process to the back of the queue), which
empties the list between the check and the retrieval.

Fixes: feenkcom/gtoolkit#1635


https://github.com/feenkcom/Brick/commit/c0a9f5 by Aliaksei Syrel
add itemId to the list

https://github.com/feenkcom/gtoolkit-coder/commit/617947 by Aliaksei Syrel
add unique id to the coder model

https://github.com/feenkcom/gtoolkit-coder/commit/8d49d6 by Aliaksei Syrel
make sure that text modified event is sent when source code is replaced

https://github.com/feenkcom/gtoolkit-coder/commit/bd4315 by Aliaksei Syrel
simplify coder actions (element is not needed)

https://github.com/feenkcom/gtoolkit-coder/commit/3df815 by Aliaksei Syrel
use #save instead of #saveFrom:

https://github.com/feenkcom/gtoolkit-coder/commit/0a9569 by Aliaksei Syrel
Merge 6746219c72af955128aee340b43dace1a2bca5c2

https://github.com/feenkcom/gtoolkit-coder/commit/f76512 by Aliaksei Syrel
[feenkcom/gtoolkit#1618] working on the method coder

https://github.com/feenkcom/gtoolkit-debugger/commit/605d79 by Aliaksei Syrel
Merge 36dfc78e204a0c6eb75b449b4747ef6c68dbe7ca

https://github.com/feenkcom/gtoolkit-debugger/commit/ad8ae6 by Aliaksei Syrel
use coder ui models in the stack list