Release version v1.0.110
Metacello new
    baseline: 'GToolkitForPharo9';
    repository: 'github://feenkcom/gtoolkit:v1.0.110/src';
    load

All commits (including upstream repositories) since last build:
https://github.com/feenkcom/gt4changes/commit/34c350 by Andrei ChiÈ
Fix the position of the dropdown for appling changes.

https://github.com/feenkcom/gt4gemstone/commit/55bd32 by Andrei Chis
Fix the dropdown for showing the list of sessions when creating a new session.

https://github.com/feenkcom/gt4gemstone/commit/159108 by Andrei Chis
Show the session name for GemStone actions in coder

https://github.com/feenkcom/gt4gemstone/commit/d69e7b by Andrei Chis
Pass on the various UIs for showing GemStone sessions