Release version v1.0.408
Metacello new
    baseline: 'GToolkitForPharo9';
    repository: 'github://feenkcom/gtoolkit:v1.0.408/src';
    load

All commits (including upstream repositories) since last build:
https://github.com/feenkcom/Bloc/commit/57ff32 by Aliaksei Syrel
prevent zero divide in BlElementAsFormFuture

https://github.com/feenkcom/zinc/commit/689541 by Sven Van Caekenberghe
Merge pull request #12 from svenvc/master

Update zinc: fix edge case in ZnUTF8Encoder>>#ensureAtBeginOfCodePointOnStream:

https://github.com/feenkcom/zinc/commit/0ad0a6 by Sven Van Caekenberghe
Merge pull request #132 from shoshanatech/wrapped-stream-raw-stream

Add #rawStream, analogous to #wrappedStream but accessing the innermost wrapped stream

https://github.com/feenkcom/zinc/commit/952dc0 by Daniel Slomovits
Add #rawStream, answering the innermost wrapped stream

Make sure SocketStream also understands #wrappedStream, and now #rawStream, as this is another possible "base stream" class.

https://github.com/feenkcom/zinc/commit/3c4fb7 by Daniel Slomovits
Add Stream>>wrappedStream returning nil so all stream-like things understand it

https://github.com/feenkcom/zinc/commit/41b613 by Daniel Slomovits
Add ZnBufferedWriteStream>>wrappedStream, all other Zinc streams understand it.

https://github.com/feenkcom/zinc/commit/580eb6 by Sven Van Caekenberghe
Merge pull request #131 from shoshanatech/explicit-position-to-end

Don't check for being in the middle of a codepoint when at end-of-stream (fixes #130)

https://github.com/feenkcom/zinc/commit/6882f5 by Daniel Slomovits
Don't check for being in the middle of a codepoint when at end-of-stream (fixes #130)

If we #position: a character readStream and the argument corresponds to end-of-stream, we don't want to try to read off the end of the stream--but we can be sure that we're not in the middle of a (valid) codepoint, regardless.