Release version v1.0.1029
Metacello new
    baseline: 'GToolkitForPharo9';
    repository: 'github://feenkcom/gtoolkit:v1.0.1029/src';
    load

All commits (including upstream repositories) since last build:
https://github.com/feenkcom/gt4pharo/commit/2ce93f by Alistair Grant
[feenkcom/gtoolkit#3934] GtVmMemoryManagement class>>validateMinOldSpaceForGc: handle maxOldSpaceSize = 0

https://github.com/feenkcom/gt4pharo/commit/5886c3 by Alistair Grant
GtVmMemoryManagementExamples to test activeDuring: return value.

https://github.com/feenkcom/gt4pharo/commit/05d169 by Alistair Grant
GtVmMemoryManagement>>activeDuring: answer the result of the supplied block

https://github.com/feenkcom/gt4pharo/commit/5f6b50 by Alistair Grant
[feenkcom/gtoolkit#3934]  GtVmMemoryManagement typos


https://github.com/feenkcom/gt4pharo/commit/d317f1 by Alistair Grant
[feenkcom/gtoolkit#3934] Add GtVmMemoryManagement.

Extend the base functionality for managing object space in the VM.

This adds minOldSpaceForGc, which specifies the minimum size old space should grow to before garbage collecting. In combination with the fullGcRatio, this can allow code that would otherwise trigger a large number of full GCs to complete with a single full GC at the end.

This commit adds the functionality, but is not yet enabled.


https://github.com/feenkcom/zinc/commit/87efcb by Sven Van Caekenberghe
Added ZnProxyServerDelegate[Test]
This is an example of a simple proxy server
Thanks Nobert Hartl for the idea


https://github.com/feenkcom/gtoolkit-remoterunner/commit/3b9600 by Alistair Grant
GtRrWorkerCountCalculatorExamples update for most recent GtRrWorkerCountImageMemoryModel enhancements

https://github.com/feenkcom/gtoolkit-remoterunner/commit/bf2010 by Alistair Grant
GtRrWorkerCountImageMemoryModel: extend with additionalOldSpace.

This handles the scenario where a known amount of old-space memory is required, but the on-disk image size can vary a lot.