Release version v1.0.1859
Metacello new
    baseline: 'GToolkitForPharo9';
    repository: 'github://feenkcom/gtoolkit:v1.0.1859/src';
    load

All commits (including upstream repositories) since last build:
https://github.com/feenkcom/gt4llm/commit/cbd224 by Tudor Girba
only show diff and code views for the assistant role. do not show the action view in the chat because we mostly have it in the header. the header of a chat message should not spawn the action because it is most often not needed there feenkcom/gtoolkit#4462